Refactored calibration screen components
